requirements.txt
# FastAPI backend
fastapi==0.101.0
uvicorn[standard]==0.23.2
pydantic==2.6.0

# Streamlit frontend
streamlit==1.27.0
pandas==2.1.0
requests==2.31.0

# PDF processing
PyMuPDF==1.23.0

# FAISS for vector storage
faiss-cpu==1.7.4

# OpenAI client
openai==1.32.0

# Environment variables
python-dotenv==1.1.1
